# schemas/openapi.yaml
openapi: "3.1.0"
info:
  title: Stigmergic Blackboard Protocol (SBP)
  version: "0.2.0"
  description: |
    REST/JSON-RPC API for the Stigmergic Blackboard Protocol.
    SBP enables decoupled, self-organizing multi-agent coordination through
    environment-based signaling with decaying pheromones and durable traces.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: SBP Maintainers
    url: https://github.com/advicenxt/sbp
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /sbp:
    post:
      operationId: jsonRpcPost
      summary: Send a JSON-RPC request
      description: |
        All SBP operations (emit, sniff, register_scent, deregister_scent,
        evaporate, inspect) are dispatched via JSON-RPC 2.0 over this endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonRpcRequest"
            examples:
              emit:
                summary: Emit a pheromone
                value:
                  jsonrpc: "2.0"
                  id: "req-1"
                  method: "sbp/emit"
                  params:
                    trail: "market.signals"
                    type: "volatility"
                    intensity: 0.8
                    decay: { type: "exponential", half_life_ms: 300000 }
                    payload: { symbol: "BTC-USD", vix_equivalent: 45.2 }
                    tags: ["crypto", "high-priority"]
                    merge_strategy: "reinforce"
              sniff:
                summary: Sniff for pheromones
                value:
                  jsonrpc: "2.0"
                  id: "req-2"
                  method: "sbp/sniff"
                  params:
                    trails: ["market.signals"]
                    min_intensity: 0.1
                    limit: 50
              register_scent:
                summary: Register a scent condition
                value:
                  jsonrpc: "2.0"
                  id: "req-3"
                  method: "sbp/register_scent"
                  params:
                    scent_id: "volatility-watcher"
                    agent_endpoint: "https://agents.example.com/handler"
                    condition:
                      type: "threshold"
                      trail: "market.signals"
                      signal_type: "volatility"
                      aggregation: "max"
                      operator: ">="
                      value: 0.7
                    cooldown_ms: 60000
      responses:
        "200":
          description: JSON-RPC response (success or error)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/JsonRpcSuccessResponse"
                  - $ref: "#/components/schemas/JsonRpcErrorResponse"
        "400":
          description: Malformed request (not valid JSON)
        "401":
          description: Missing or invalid API key
        "429":
          description: Rate limited

    get:
      operationId: sseStream
      summary: Open an SSE stream for triggers
      description: |
        Opens a Server-Sent Events stream to receive scent trigger notifications
        in real-time. Each event contains a JSON-RPC notification.
      parameters:
        - name: Sbp-Session-Id
          in: header
          required: false
          schema:
            type: string
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
          description: Last received event ID for resumability
      responses:
        "200":
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Server-Sent Events stream. Each event has:
                  - event: "message"
                  - id: unique event ID
                  - data: JSON-RPC trigger notification
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    JsonRpcRequest:
      type: object
      required: [jsonrpc, id, method, params]
      properties:
        jsonrpc:
          type: string
          const: "2.0"
        id:
          oneOf:
            - type: string
            - type: integer
        method:
          type: string
          enum:
            - sbp/emit
            - sbp/sniff
            - sbp/register_scent
            - sbp/deregister_scent
            - sbp/evaporate
            - sbp/inspect
            - sbp/inscribe
            - sbp/read
            - sbp/erase
        params:
          type: object
    JsonRpcSuccessResponse:
      type: object
      required: [jsonrpc, id, result]
      properties:
        jsonrpc:
          type: string
          const: "2.0"
        id:
          oneOf:
            - type: string
            - type: integer
        result:
          type: object
    JsonRpcErrorResponse:
      type: object
      required: [jsonrpc, id, error]
      properties:
        jsonrpc:
          type: string
          const: "2.0"
        id:
          oneOf:
            - type: string
            - type: integer
            - type: "null"
        error:
          $ref: "#/components/schemas/JsonRpcError"
    JsonRpcError:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: |
            SBP error codes:
            -32700: Parse error
            -32600: Invalid request
            -32601: Method not found
            -32602: Invalid params
            -32001: Trail not found
            -32002: Scent not found
            -32003: Payload validation failed
            -32004: Rate limited
            -32005: Unauthorized
            -32007: Trace not found
            -32008: Trace value too large
        message:
          type: string
        data:
          description: Additional error context

    # -- Decay Models --
...
security:
  - ApiKeyAuth: []

    # -- Trace --
    Trace:
      type: object
      required: [id, trail, key, value, created_at, updated_at, version, tags]
      properties:
        id:
          type: string
        trail:
          type: string
        key:
          type: string
        value:
          type: object
          description: Arbitrary JSON payload (max 1 MB)
        created_at:
          type: integer
        updated_at:
          type: integer
        version:
          type: integer
        source_agent:
          type: string
        tags:
          type: array
          items:
            type: string

    InscribeParams:
      type: object
      required: [trail, key, value]
      properties:
        trail:
          type: string
          minLength: 1
        key:
          type: string
          minLength: 1
        value:
...